Example: Union Types
https://dotty.epfl.ch/docs/reference/new-types/union-types.html
